basic subclassing
